requirements.txt
transformers==4.44.2
accelerate==0.34.2
diffusers==0.30.2
fastapi[standards]==0.114.0
uvicorn==0.30.6
python-dotenv==1.0.1
python-multipart==0.0.9